wip: admin Code refactor to seperate files
# Conflicts:
#	tests/setup.spec.ts
