--- tests/poms/adminhtml/magentoAdmin.page.ts
@@ -1,340 +1,0 @@
-// @ts-check
-
-import { expect, type Locator, type Page } from '@playwright/test';
-import { UIReference, inputValues, outcomeMarker } from '@config';
-import { requireEnv } from '@utils/env.utils';
-
-class MagentoAdminPage {
-  readonly page: Page;
-  readonly adminLoginEmailField: Locator;
-  readonly adminLoginPasswordField: Locator;
-  readonly adminLoginButton: Locator;
-
-  constructor(page: Page) {
-    this.page = page;
-    this.adminLoginEmailField = page.getByLabel(UIReference.magentoAdminPage.usernameFieldLabel);
-    this.adminLoginPasswordField = page.getByLabel(UIReference.magentoAdminPage.passwordFieldLabel);
-    this.adminLoginButton = page.getByRole('button', {name: UIReference.magentoAdminPage.loginButtonLabel});
-  }
-
-  /**
-   * @feature Cart Price Rules Configuration
-   * @scenario Add or activate a cart price rule with a specific coupon code
-   * @given the admin is on the Magento dashboard
-   * @when the coupon exists but is inactive
-   * @then the admin activates the existing coupon and saves the rule
-   * @but if the coupon does not exist
-   * @then the admin creates a new cart price rule with the given coupon code
-   * @and selects all websites and customer groups
-   * @and sets the coupon type and discount amount
-   * @and clicks the Save button
-   * @then the system displays a success message confirming the rule was saved
-   */
-  async addCartPriceRule(magentoCouponCode: string){
-    let resultMessage = "";
-
-    // Force specific viewport size to deal with webkit issues
-    await this.page.setViewportSize({
-        width: 1920,
-        height: 1080
-    })
-
-    const mainMenuMarketingButton = this.page.getByRole('link', {name: UIReference.magentoAdminPage.navigation.marketingButtonLabel});
-    const cartPriceRulesLink = this.page.getByRole('link', {name: UIReference.magentoAdminPage.subNavigation.cartPriceRulesButtonLabel});
-    await expect(mainMenuMarketingButton, `Button for Marketing is visible`).toBeVisible();
-
-    await expect(async () => {
-      await mainMenuMarketingButton.click();
-      await expect(cartPriceRulesLink, `Button for Cart Price Rules is visible`).toBeVisible();
-    }).toPass();
-
-    await cartPriceRulesLink.click();
-
-    const addCartPriceRuleButton = this.page.getByRole('button', {name: UIReference.cartPriceRulesPage.addCartPriceRuleButtonLabel});
-    await addCartPriceRuleButton.waitFor();
-
-    // Use search field to check for coupon codes.
-    const couponSearchField = this.page.locator(UIReference.magentoAdminPage.couponSearchFieldLocator);
-    await couponSearchField.fill(magentoCouponCode);
-    await this.page.getByRole('button', {name: UIReference.general.searchButtonLabel}).click();
-
-    await expect(this.page.getByText(UIReference.magentoAdminPage.searchResultsText), `Search results text visible`).toBeVisible();
-
-    const couponCellField = this.page.getByRole('cell', { name: outcomeMarker.magentoAdmin.noResultsFoundText });
-
-    if(await couponCellField.isHidden()){
-      const couponStatusField = this.page.locator('tr').filter({hasText:magentoCouponCode}).first();
-      const couponStatus = await couponStatusField.innerText();
-      if(couponStatus.includes(UIReference.cartPriceRulesPage.couponCodeActiveStatusText)){
-          resultMessage = 'Coupon already exists and is active.';
-      } else {
-        // coupon has been found, but is not active.
-        await couponStatusField.click();
-        const activeStatusSwitcher = this.page.locator(UIReference.cartPriceRulesPage.activeStatusSwitcherLocator).first();
-        const activeStatusLabel = this.page.locator(UIReference.cartPriceRulesPage.activeStatusLabelLocator).first();
-
-        await expect(activeStatusLabel, `Active/Disable toggle is visible`).toBeVisible();
-        await activeStatusSwitcher.click();
-
-        const saveCouponButton = this.page.getByRole('button', {name:UIReference.cartPriceRulesPage.saveRuleButtonLabel, exact:true});
-        await saveCouponButton.click();
-
-        await expect(this.page.locator(
-          UIReference.general.messageLocator).filter({hasText: outcomeMarker.magentoAdmin.couponRuleSavedText}
-        ), "Message 'you saved the rule' is visible").toBeVisible();
-        resultMessage = `Coupon code ${magentoCouponCode} has been activated.`;
-      }
-    } else {
-      // coupon is not set
-      await addCartPriceRuleButton.click();
-
-      const websiteSelector = this.page.getByLabel(UIReference.cartPriceRulesPage.websitesSelectLabel);
-      await websiteSelector.evaluate(select => {
-        const s = select as HTMLSelectElement;
-        for (const option of s.options) {
-          option.selected = true;
-        }
-        select.dispatchEvent(new Event('change'));
-      });
-
-      const customerGroupsSelector = this.page.getByLabel(UIReference.cartPriceRulesPage.customerGroupsSelectLabel, { exact: true });
-      await customerGroupsSelector.evaluate(select => {
-        const s = select as HTMLSelectElement;
-        for (const option of s.options) {
-          option.selected = true;
-        }
-        select.dispatchEvent(new Event('change'));
-      });
-
-      await this.page.getByRole('textbox', { name: UIReference.cartPriceRulesPage.ruleNameFieldLabel }).fill(magentoCouponCode);
-      await this.page.locator(UIReference.cartPriceRulesPage.couponTypeSelectField).selectOption({ label: inputValues.coupon.couponType });
-      await this.page.getByLabel(UIReference.cartPriceRulesPage.couponCodeFieldLabel).fill(magentoCouponCode);
-
-      await this.page.getByText(UIReference.cartPriceRulesPage.actionsSubtitleLabel, { exact: true }).click();
-      await this.page.getByLabel(UIReference.cartPriceRulesPage.discountAmountFieldLabel).fill('10');
-
-      const couponSaveButton = this.page.getByRole('button', { name: UIReference.cartPriceRulesPage.saveRuleButtonLabel, exact: true });
-      await couponSaveButton.scrollIntoViewIfNeeded();
-      await couponSaveButton.click({force:true});
-      await expect(this.page.locator(
-          UIReference.general.messageLocator).filter({hasText: outcomeMarker.magentoAdmin.couponRuleSavedText}
-      ), "Message 'you saved the rule' is visible").toBeVisible();
-      resultMessage = `Coupon code ${magentoCouponCode} has been set and activated.`;
-    }
-
-    // Clear the search field
-    await couponSearchField.waitFor();
-    const clearSearchButton = this.page.getByRole('button', { name: UIReference.cartPriceRulesPage.clearSearchButtonLabel });
-    await clearSearchButton.click();
-    await expect(couponSearchField, `Coupon Code search field is empty`).toBeEmpty();
-
-    return resultMessage;
-  };
-
-  /**
-   * @feature Customer Management
-   * @scenario Check if a customer exists by email address
-   * @given the admin is on the Magento dashboard
-   * @when the admin navigates to Customers > All Customers
-   * @and the customer table is fully loaded
-   * @and the admin searches for a specific email address
-   * @then the system returns whether a customer with that email exists in the customer list
-   */
-  async checkIfCustomerExists(email: string){
-    const mainMenuCustomersButton = this.page.getByRole('link', {name: UIReference.magentoAdminPage.navigation.customersButtonLabel}).first();
-    const allCustomersLink = this.page.getByRole('link', {name: UIReference.magentoAdminPage.subNavigation.allCustomersButtonLabel});
-    const customersSearchField = this.page.getByRole('textbox', {name: UIReference.customerOverviewPage.tableSearchFieldLabel});
-
-    // loop clicking the 'Customers' button until clicking it show the subnavigation
-    await expect(async() =>{
-      await mainMenuCustomersButton.press('Enter');
-      await expect(allCustomersLink, `Link to "All Customers" is visible`).toBeVisible({timeout: 5000});
-    }).toPass();
-
-    await allCustomersLink.click();
-
-    // Wait for URL. If loading symbol is visible, wait for it to go away
-    await this.page.waitForURL(`**/${requireEnv('MAGENTO_ADMIN_SLUG')}/customer/index/**`);
-    if (await this.page.locator(UIReference.general.loadingSpinnerLocator).isVisible()) {
-      await this.page.locator(UIReference.general.loadingSpinnerLocator).waitFor({state: 'hidden'});
-    }
-
-    await customersSearchField.waitFor();
-    await customersSearchField.fill(email);
-    await this.page.getByRole('button', {name: UIReference.general.searchButtonLabel}).click();
-
-
-    if (await this.page.locator(UIReference.general.loadingSpinnerLocator).isVisible()) {
-      await this.page.locator(UIReference.general.loadingSpinnerLocator).waitFor({state: 'hidden'});
-    }
-
-    // Loop to ensure the 'results found' text is visible
-    await expect(async() =>{
-      await this.page.getByText(outcomeMarker.customerOverviewPage.searchResultsFoundText).first();
-    }).toPass();
-
-    // Return true (email found) or false (email not found)
-    return await this.page.getByRole('cell', {name:email}).locator('div').isVisible();
-  }
-
-  /**
-   * @feature Magento Admin Configuration
-   * @scenario Disable the login CAPTCHA on the admin panel
-   * @given the admin is logged into the Magento dashboard
-   * @when the admin navigates to Stores > Configuration > Customers > Customer Configuration > CAPTCHA section
-   * @and the "Use system value" checkbox for CAPTCHA is unchecked
-   * @and the "Enable CAPTCHA on Admin Login" select field is visible
-   * @and the current setting is "Yes"
-   * @then the admin changes the setting to "No"
-   * @and clicks the Save Config button
-   * @then the system displays a success message confirming the configuration was saved
-   */
-  async disableLoginCaptcha() {
-    const mainMenuStoresButton = this.page.getByRole('link', { name: UIReference.magentoAdminPage.navigation.storesButtonLabel});
-    // selecting first specifically because plugins can place another 'configuration' link in this menu.
-    const storeSettingsConfigurationLink = this.page.getByRole('link', { name: UIReference.magentoAdminPage.subNavigation.configurationButtonLabel }).first();
-
-    await expect(async () => {
-      await mainMenuStoresButton.click();
-      await expect(storeSettingsConfigurationLink, `Link to Store Configuration is visible`).toBeVisible();
-    }).toPass();
-
-    await storeSettingsConfigurationLink.click();
-
-    const customersTab = this.page.getByRole('tab', { name: UIReference.configurationPage.customersTabLabel });
-    const customerConfigurationLink = this.page.getByRole('link', { name: UIReference.configurationPage.customerConfigurationTabLabel });
-    await customersTab.click();
-    await customerConfigurationLink.waitFor();
-    await customerConfigurationLink.click();
-
-    const captchaSettingsBlock = this.page.getByRole('link', { name: UIReference.configurationPage.captchaSectionLabel })
-      .filter({hasNotText: 'documentation'});
-    const captchaSettingsSystemValueCheckbox = this.page.locator(UIReference.configurationPage.captchaSettingSystemCheckbox);
-
-    await captchaSettingsBlock.waitFor();
-
-    if(!await captchaSettingsSystemValueCheckbox.isVisible()) {
-      await captchaSettingsBlock.click();
-    }
-
-    if(await captchaSettingsSystemValueCheckbox.isChecked()){
-      await captchaSettingsSystemValueCheckbox.uncheck();
-    }
-
-    const captchaSettingSelectField = this.page.locator(UIReference.configurationPage.captchaSettingSelectField);
-    const selectedOption = await captchaSettingSelectField.locator('option:checked').textContent();
-
-    // We only have to perform these steps if the option is set to 'Yes'
-    if(selectedOption == 'Yes') {
-      await captchaSettingSelectField.selectOption({label: inputValues.captcha.captchaDisabled});
-
-      const saveConfigButton = this.page.getByRole('button', { name: UIReference.configurationPage.saveConfigButtonLabel });
-      await saveConfigButton.click();
-
-      await expect(this.page.locator(UIReference.general.messageLocator).filter(
-        {hasText: outcomeMarker.magentoAdmin.configurationSavedText}),
-        `Notification "${outcomeMarker.magentoAdmin.configurationSavedText}" is visible`).toBeVisible();
-    }
-  }
-
-  /**
-   * @feature Enable multiple admin logins in Magento
-   * @scenario Admin enables the ability for multiple users to log in with the same admin account
-   * @given the user is on the Magento admin dashboard
-   * @when the user navigates to Stores > Configuration > Advanced > Admin > Security
-   * @and the "Allow Multiple Admin Account Login" field is visible
-   * @and the "Use system value" checkbox is unchecked
-   * @and the select field value is "No"
-   * @then the user selects "Yes" from the dropdown
-   * @and clicks the Save Config button
-   * @then the system displays a success message
-   */
-  async enableMultipleAdminLogins() {
-    const mainMenuStoresButton = this.page.getByRole('link', { name: UIReference.magentoAdminPage.navigation.storesButtonLabel});
-    // selecting first specifically because plugins can place another 'configuration' link in this menu.
-    const storeSettingsConfigurationLink = this.page.getByRole('link', { name: UIReference.magentoAdminPage.subNavigation.configurationButtonLabel }).first();
-
-    await expect(async () => {
-      await mainMenuStoresButton.click();
-      await expect(storeSettingsConfigurationLink, `Link to Store Configuration is visible`).toBeVisible();
-    }).toPass();
-
-    await storeSettingsConfigurationLink.click();
-
-    const advancedConfigurationTab = this.page.getByRole('tab', { name: UIReference.configurationPage.advancedTabLabel });
-    const advancedConfigAdminLabel = this.page.getByRole('link', { name: UIReference.configurationPage.advancedAdministrationTabLabel, exact: true });
-    await advancedConfigurationTab.click();
-    await advancedConfigAdminLabel.waitFor();
-    await advancedConfigAdminLabel.click();
-
-    const advancedConfigSecuritySection = this.page.getByRole('link', { name: UIReference.configurationPage.securitySectionLabel });
-    const multipleLoginsSystemCheckbox = this.page.locator(UIReference.configurationPage.allowMultipleLoginsSystemCheckbox);
-
-    await advancedConfigSecuritySection.waitFor();
-    if (!await multipleLoginsSystemCheckbox.isVisible()) {
-      await advancedConfigSecuritySection.click();
-    }
-
-    await expect(multipleLoginsSystemCheckbox, `Checkbox for multiple admin logins is visible`).toBeVisible();
-
-    // make sure the 'use system value' option is not checked
-    const adminAccountSharingSystemValueCheckbox = this.page.locator(UIReference.configurationPage.allowMultipleLoginsSystemCheckbox);
-    if (await adminAccountSharingSystemValueCheckbox.isChecked()) {
-      await adminAccountSharingSystemValueCheckbox.uncheck();
-    }
-
-    const allowMultipleLoginSelectField = this.page.locator(UIReference.configurationPage.allowMultipleLoginsSelectField);
-    const selectedOption = await allowMultipleLoginSelectField.locator('option:checked').textContent();
-
-    // We only have to perform these steps if the option is set to 'No'
-    if(selectedOption == 'No') {
-      await allowMultipleLoginSelectField.selectOption({label: inputValues.adminLogins.allowMultipleLogins});
-
-      const saveConfigButton = this.page.getByRole('button', { name: UIReference.configurationPage.saveConfigButtonLabel });
-      await saveConfigButton.click();
-
-      await expect(this.page.locator(UIReference.general.messageLocator).filter(
-        {hasText: outcomeMarker.magentoAdmin.configurationSavedText}),
-        `Notification "${outcomeMarker.magentoAdmin.configurationSavedText}" is visible`).toBeVisible();
-    }
-  }
-
-  /**
-   * @feature Login to Magento admin dashboard
-   * @scenario User logs in to admin dashboard
-   * @given the admin slug environment variable is defined
-   * @and the user navigates to the admin login page
-   * @when the user enters a valid username and password
-   * @and the user clicks the login button
-   * @then the user should see the dashboard heading displayed
-   */
-  async login(username: string, password: string){
-    await this.page.goto(requireEnv('MAGENTO_ADMIN_SLUG'));
-    await this.page.waitForURL(`**/${requireEnv('MAGENTO_ADMIN_SLUG')}`);
-
-    if(await this.page.getByRole('heading', {name: UIReference.magentoAdminPage.dashboardHeadingText}).isVisible()) {
-      // already logged in
-      return;
-    }
-
-    await this.adminLoginEmailField.fill(username);
-    await this.adminLoginPasswordField.fill(password);
-    await this.adminLoginButton.click();
-
-    const captchaNotification = this.page.locator(UIReference.general.messageLocator).filter({hasText: UIReference.magentoAdminPage.captchaIncorrectText});
-
-    if(await captchaNotification.isVisible()) {
-      console.log('CAPTCHA field is visible, automated login not possible!');
-      throw new Error("CAPTCHA field is visible, automated login not possible!");
-    }
-
-    const dashboardLabel = this.page.getByRole('heading',{level:1, name: UIReference.magentoAdminPage.dashboardHeadingText});
-
-    // expect the H1 'Dashboard' to be visible
-    await expect(async () => {
-      await expect(dashboardLabel, `Title "Dashboard" is visible`).toBeVisible();
-    }).toPass();
-  }
-}
-
-export default MagentoAdminPage;
